+++ packages/utils/test/schema/getDefaultFormStateTest.ts
@@ -1342,6 +1342,28 @@ export default function getDefaultFormStateTest(testValidator: TestValidatorType
         });
       });
 
+      describe('an object with a valid formData and enum property with default value', () => {
+        test('getDefaultFormState', () => {
+          const schema: RJSFSchema = {
+            type: 'object',
+            properties: {
+              test: {
+                type: 'string',
+                enum: [
+                  { label: 'a', value: 'a' },
+                  { label: 'b', value: 'b' },
+                ],
+                default: { label: 'a', value: 'a' },
+              },
+            },
+          };
+
+          expect(getDefaultFormState(testValidator, schema, { test: { label: 'b', value: 'b' } })).toEqual({
+            test: { label: 'b', value: 'b' },
+          });
+        });
+      });
+
       describe('oneOf with const values', () => {
         const schema: RJSFSchema = {
           type: 'object',
@@ -1969,6 +1991,22 @@ export default function getDefaultFormStateTest(testValidator: TestValidatorType
         };
 
         expect(ensureFormDataMatchingSchema(testValidator, schema, schema, 'a')).toEqual('a');
+      });
+
+      it('Test schema with valid formData with an enum and its default value', () => {
+        schema = {
+          type: 'string',
+          enum: [
+            { label: 'a', value: 'a' },
+            { label: 'b', value: 'b' },
+          ],
+          default: { label: 'a', value: 'a' },
+        };
+
+        expect(getDefaultFormState(testValidator, schema, { label: 'b', value: 'b' })).toEqual({
+          label: 'b',
+          value: 'b',
+        });
       });
     });
     describe('AJV $data reference in const property in schema should not be treated as default/const value', () => {